Move farmer policy to user folder
