+++ app/Providers/AuthServiceProvider.php
@@ -10,12 +10,12 @@ use App\Models\Site\Province;
 use App\Models\Site\Region;
 use App\Models\User\Farmer\Farmer;
 use App\Policies\BatchPolicy;
-use App\Policies\FarmerPolicy;
 use App\Policies\FarmerReportPolicy;
 use App\Policies\FarmlandPolicy;
 use App\Policies\Site\MunicityPolicy;
 use App\Policies\Site\ProvincePolicy;
 use App\Policies\Site\RegionPolicy;
+use App\Policies\User\FarmerPolicy;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 
 class AuthServiceProvider extends ServiceProvider
